pyproject.toml
[project]
name = "gemini-mcp-embedding-server"
version = "0.1.0"
description = "MCP server connecting local files to Claude via Gemini Embedding 2"
authors = [
    { name = "Alaeddine Messadi" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-genai>=0.2.2",
    "chromadb>=0.4.24",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0"
]

[project.scripts]
gemini-mcp = "gemini_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_mcp"]
